Refactor Nav into its own feature
